pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kao-legal"
version = "0.1.0"
description = "Legal Advisor Agent SaaS with Strands Agents SDK and Amazon Bedrock"
requires-python = ">=3.12"
authors = [
    {name = "Kao Legal Team"}
]
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "boto3>=1.34.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "mangum>=0.19.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "langfuse>=3.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "structlog>=24.4.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "stripe>=9.0.0",
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "moto[s3,dynamodb,comprehend]>=5.0.0",
    "pytest-mock>=3.14.0",
    "boto3-stubs[s3,dynamodb,comprehend,bedrock-runtime]>=1.34.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "W"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src"]
markers = ["integration: marks tests requiring AWS credentials"]
